requirements.txt
# Django
Django

# Django REST Framework
djangorestframework

# Testing
pytest
pytest-django
beautifulsoup4
coverage